refactor classes in data directory into utils
