+++ src/main/java/com/github/mkram17/bazaarutils/utils/storage/UserOrdersStorage.java
@@ -1,7 +1,6 @@
-package com.github.mkram17.bazaarutils.data;
+package com.github.mkram17.bazaarutils.utils.storage;
 
 import com.github.mkram17.bazaarutils.utils.bazaar.market.order.Order;
-import com.github.mkram17.bazaarutils.utils.storage.DataStorage;
 import com.google.gson.reflect.TypeToken;
 import java.lang.reflect.Type;
 import java.util.ArrayList;
